Generate pages and feeds for tags
The code has been adapted from the categories plugin. It isn't a
complete clone as I couldn't understand the code over there ( 2
months ago, when I wrote this ) I think it's best to commit this
code, and then try to make it similar to the categories plugin.

Maybe the two can be combined.